requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
python-dotenv>=1.0
openai>=1.14
requests>=2.32
sentence-transformers>=2.6
faiss-cpu>=1.7
streamlit>=1.34